Added Date and Time to log file
Made it so that the date and times are attached to the log file so that a new file will be created each time a new session is started.  Added Boolean to enable the user to have the ability to turn on /off the debug so they don't fill up their server with logs. Default is false.